integrate database

+++ database/src/main/java/com/mshdabiola/database/model/PawnEntity.kt
@@ -7,7 +7,7 @@ import com.mshdabiola.ludo.model.Pawn
 @Entity(tableName = "pawn_table", primaryKeys = ["id","gameId"])
 data class PawnEntity(
     val id: Int = 0, //index is the id
-    val gameId : Int,
+    val gameId : Long,
     val currentPos: Int,
     val playerId :Int
 )
@@ -17,6 +17,6 @@ fun Int.toColor()= GameColor.values()[this/4]
 
 fun Int.toId()=(this%4)+1
 
-fun Pawn.toPawnEntity(playerId: Int,gameId: Int)=PawnEntity(this.index,gameId,currentPos,playerId)
+fun Pawn.toPawnEntity(playerId: Int,gameId: Long)=PawnEntity(this.index,gameId,currentPos,playerId)
 
 fun PawnEntity.toPawn()=Pawn().copy(id.toId(),currentPos,id.toColor())
